backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
psycopg2-binary==2.9.3 